be more tolerant on PACKAGE_NAME argument
when a package name is given to `create-xdc.sh`,
this may be just a copy+pasted filename.
that, however, results in double `.xdc` extensions.

this pr accepts PACKAGE_NAME with or without .xdc extensions,
of course, when creating the final ZIP, .xdc is always added,
however only once :)

moreover, this pr documents the weird string manipulation commands.