make repo easier useable as a template
this repo is meant as a super-lightweight demo for creating .xdc,
gist is to clone this repo and to change its content.

the `images` directory was a bit in the way of that as
it would be either included to the `.xdc`
or the `create-xdc.sh` would need and additional entry -
doable, but that would make the example a bit less smart.

this pr solves the issue by just deleting the image -
it was not really needed, it just repeats what is shown in the text above -
and it also easily get outdated.
also, this image created a bit of unbalanced - it is the only image in the README
and shows only a very small aspect of a smaller browser.

if we want to add more images to the README, we have to rethink that
(could we put them to `.github?`)
but for now it seems fine?