use graphical icon that not just repeats the name
the old icon just uses the text "hello" black on white.

while maybe stylish,
set in context where the icon is usually used,
this looks repetitive and a bit weird,
esp. as the icon has also white background.

of course, this is not "wrong" and other apps can pick up this style,
but we should not use that as a default pattern we encourage ppl to use.

starburst inspired from https://codepen.io/lethargic/pen/aBybeK ,
the hand is just an emoji,
put together with gimp - if someone want to do a vector or has another idea,
this is very welcome :)